+++ packages/browser-wallet/src/shared/storage/types.ts
@@ -203,6 +203,7 @@ export interface TokenMetadata {
     thumbnail?: MetadataUrl;
     display?: MetadataUrl;
     artifact?: MetadataUrl;
+    unique?: boolean;
 }
 
 export interface TokenIdAndMetadata {