+++ packages/browser-wallet/src/shared/utils/token-helpers.ts
@@ -1,5 +1,6 @@
 import { Buffer } from 'buffer/';
-import { InstanceInfo, JsonRpcClient } from '@concordium/web-sdk';
+import { AccountAddress, InstanceInfo, JsonRpcClient } from '@concordium/web-sdk';
+import uleb128 from 'leb128/unsigned';
 import { TokenMetadata } from '@shared/storage/types';
 
 export interface ContractDetails {
@@ -37,7 +38,7 @@ export function getMetadataParameter(id: string): Buffer {
 function deserializeTokenMetadataReturnValue(returnValue: string) {
     const bufferStream = Buffer.from(returnValue, 'hex');
     const length = bufferStream.readUInt16LE(2);
-    const url = bufferStream.slice(4, 4 + length).toString('utf8');
+    const url = Buffer.from(bufferStream.subarray(4, 4 + length)).toString('utf8');
     return url;
 }
 
@@ -122,3 +123,87 @@ export async function getTokenMetadata(tokenUrl: string): Promise<TokenMetadata>
     }
     return resp.json();
 }
+
+/**
+ * Serialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#id3
+ */
+const serializeBalanceParameter = (tokenIds: string[], accountAddress: string) => {
+    const queries = Buffer.alloc(2);
+    queries.writeUInt16LE(tokenIds.length, 0);
+
+    const tokens = tokenIds.reduce((acc, t) => {
+        const token = Buffer.from(t, 'hex');
+        const tokenLength = Buffer.alloc(1);
+        tokenLength.writeUInt8(token.length, 0);
+
+        const addressType = Buffer.alloc(1); // Account address type
+        const address = new AccountAddress(accountAddress).decodedAddress;
+
+        return Buffer.concat([acc, tokenLength, token, addressType, address]);
+    }, queries);
+
+    return tokens;
+};
+
+/**
+ * Deserialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#response
+ */
+const deserializeBalanceAmounts = (value: string): bigint[] => {
+    const buf = Buffer.from(value, 'hex');
+    const n = buf.readUInt16LE(0);
+    let cursor = 2; // First 2 bytes hold number of token amounts included in response.
+    const amounts: bigint[] = [];
+
+    // eslint-disable-next-line no-plusplus
+    for (let i = 0; i < n; i++) {
+        const end = buf.subarray(cursor).findIndex((b) => b < 2 ** 7) + 1; // Find the first byte with most significant bit not set, signaling the last byte in the leb128 slice.
+
+        const amount = uleb128.decode(Buffer.from(buf.subarray(cursor, cursor + end)));
+        amounts.push(BigInt(amount));
+
+        cursor += end;
+    }
+
+    return amounts;
+};
+
+export type ContractBalances = Record<string, bigint>;
+
+export const getContractBalances = async (
+    client: JsonRpcClient,
+    contractIndex: string,
+    tokenIds: string[],
+    accountAddress: string
+): Promise<ContractBalances> => {
+    const index = BigInt(contractIndex);
+    const subindex = 0n;
+    const instanceInfo = await client.getInstanceInfo({ index, subindex });
+
+    if (instanceInfo === undefined) {
+        return {};
+    }
+
+    const result = await client.invokeContract({
+        contract: { index, subindex },
+        method: `${instanceInfo.name.substring(5)}.balanceOf`,
+        parameter: serializeBalanceParameter(tokenIds, accountAddress),
+    });
+
+    if (result === undefined || result.tag === 'failure' || result.returnValue === undefined) {
+        return {};
+    }
+
+    const amounts = deserializeBalanceAmounts(result.returnValue);
+
+    if (amounts.length !== tokenIds.length) {
+        throw new Error('Mismatch between length of requested tokens and token amounts in response.');
+    }
+
+    return tokenIds.reduce<ContractBalances>(
+        (acc, cur, i) => ({
+            ...acc,
+            [cur]: amounts[i],
+        }),
+        {}
+    );
+};