+++ packages/browser-wallet/src/types/leb128.d.ts
@@ -1,0 +1,3 @@
+declare module 'leb128/unsigned' {
+    export function decode(buffer: import('buffer/').Buffer): string;
+}