+++ packages/browser-wallet/CHANGELOG.md
@@ -7,6 +7,10 @@
 -   A toast is now shown when the passcode is successfully updated.
 -   Shows submitted transactions in the transaction list.
 
+### Changed
+
+-   Default view for account page is now CIS-2 token list
+
 ## 0.7.5
 
 ### Fixed